Add files via upload
Create phase calculations

+++ AD5933.h
@@ -138,8 +138,8 @@ class AD5933 {
 
         // Perform frequency sweeps
         static bool frequencySweep(int real[], int imag[], int);
-        static bool calibrate(double gain[], int phase[], int ref, int n);
-        static bool calibrate(double gain[], int phase[], int real[],
+        static bool calibrate(double gain[], float phase[], int ref, int n);
+        static bool calibrate(double gain[], float phase[], int real[],
                               int imag[], int ref, int n);
     private:
         // Private data
